separate metric class
